Upstream version 7.0.0~beta1

+++ imagery/i.class/graphics.c
@@ -1,6 +1,5 @@
 #include "globals.h"
 #include <grass/display.h>
-#include <grass/raster.h>
 #include "local_proto.h"
 
 static View *makeview(double bottom, double top, double left, double right)
@@ -91,7 +90,7 @@ int Outline_box(int top, int bottom, int left, int right)
 
 int Text_width(char *text)
 {
-    int top, bottom, left, right;
+    double top, bottom, left, right;
 
     R_get_text_box(text, &top, &bottom, &left, &right);
