+++ imagery/i.vpoints/dot.c
@@ -1,9 +1,9 @@
 #include "globals.h"
-#include <grass/raster.h>
+#include <grass/display.h>
 
 int dot(int x, int y)
 {
-    int vx[5], vy[5];
+    double vx[5], vy[5];
 
     vx[0] = x;
     vy[0] = y - dotsize;
